requirements.txt
pyyaml==6.0.1
pandas==2.1.4
selenium==4.15.2
google-generativeai==0.4.1
python-dotenv==1.0.1
chromedriver-autoinstaller==0.6.4 